updated README.md and config file
