+++ requirements.txt
@@ -1,2 +1,134 @@
+aiofiles==22.1.0
+aiohttp==3.8.4
+aiosignal==1.3.1
+aiosqlite==0.18.0
+anyio==3.6.2
+argon2-cffi==21.3.0
+argon2-cffi-bindings==21.2.0
+arrow==1.2.3
+asttokens==2.2.1
+async-timeout==4.0.2
+attrs==22.2.0
+Babel==2.12.1
+backcall==0.2.0
+beautifulsoup4==4.11.2
+bleach==6.0.0
+certifi==2022.12.7
+cffi==1.15.1
+charset-normalizer==3.1.0
+colorama==0.4.6
+comm==0.1.2
+contourpy==1.0.7
+cycler==0.11.0
+debugpy==1.6.6
+decorator==5.1.1
+defusedxml==0.7.1
+dnspython==2.3.0
+executing==1.2.0
+fastjsonschema==2.16.3
+fonttools==4.39.0
+fqdn==1.5.1
+frozenlist==1.3.3
+gitdb==4.0.10
+GitPython==3.1.31
+idna==3.4
+ipykernel==6.21.3
+ipython==8.11.0
+ipython-genutils==0.2.0
+isoduration==20.11.0
+jedi==0.18.2
+Jinja2==3.1.2
+joblib==1.2.0
+json5==0.9.11
+jsonpointer==2.3
+jsonschema==4.17.3
+jupyter-events==0.6.3
+jupyter-server-mathjax==0.2.6
+jupyter-ydoc==0.2.3
+jupyter_client==8.0.3
+jupyter_core==5.2.0
+jupyter_server==2.4.0
+jupyter_server_fileid==0.8.0
+jupyter_server_terminals==0.4.4
+jupyter_server_ydoc==0.6.1
+jupyterlab==3.6.1
+jupyterlab-git==0.41.0
+jupyterlab-pygments==0.2.2
+jupyterlab_server==2.20.0
+kiwisolver==1.4.4
+loguru==0.6.0
+MarkupSafe==2.1.2
+matplotlib==3.7.1
+matplotlib-inline==0.1.6
+mistune==2.0.5
+multidict==6.0.4
+nbclassic==0.5.3
+nbclient==0.7.2
+nbconvert==7.2.9
+nbdime==3.1.1
+nbformat==5.7.3
+nest-asyncio==1.5.6
+notebook==6.5.3
+notebook_shim==0.2.2
+numpy==1.24.2
+nvidia-cublas-cu11==11.10.3.66
+nvidia-cuda-nvrtc-cu11==11.7.99
+nvidia-cuda-runtime-cu11==11.7.99
+nvidia-cudnn-cu11==8.5.0.96
 openai==0.27.2
+packaging==23.0
+pandas==1.5.3
+pandocfilters==1.5.0
+parso==0.8.3
+pexpect==4.8.0
+pickleshare==0.7.5
+Pillow==9.4.0
 pinecone-client==2.2.1
+platformdirs==3.1.1
+plotly==5.13.1
+prometheus-client==0.16.0
+prompt-toolkit==3.0.38
+psutil==5.9.4
+ptyprocess==0.7.0
+pure-eval==0.2.2
+pycparser==2.21
+Pygments==2.14.0
+pyparsing==3.0.9
+pyrsistent==0.19.3
+python-dateutil==2.8.2
+python-dotenv==1.0.0
+python-json-logger==2.0.7
+pytz==2022.7.1
+PyYAML==6.0
+pyzmq==25.0.1
+requests==2.28.2
+rfc3339-validator==0.1.4
+rfc3986-validator==0.1.1
+scikit-learn==1.2.2
+scipy==1.10.1
+seaborn==0.12.2
+Send2Trash==1.8.0
+six==1.16.0
+smmap==5.0.0
+sniffio==1.3.0
+soupsieve==2.4
+stack-data==0.6.2
+tenacity==8.2.2
+terminado==0.17.1
+threadpoolctl==3.1.0
+tinycss2==1.2.1
+tomli==2.0.1
+torch==1.13.1
+tornado==6.2
+tqdm==4.65.0
+traitlets==5.9.0
+typing_extensions==4.5.0
+uri-template==1.2.0
+urllib3==1.26.15
+wcwidth==0.2.6
+webcolors==1.12
+webencodings==0.5.1
+websocket-client==1.5.1
+y-py==0.5.9
+yarl==1.8.2
+ypy-websocket==0.8.2
